+++ server/impl/src/main/java/com/walmartlabs/concord/server/org/project/ProjectRepositoryManager.java
@@ -22,6 +22,9 @@ package com.walmartlabs.concord.server.org.project;
 
 import com.walmartlabs.concord.project.ProjectLoader;
 import com.walmartlabs.concord.project.model.ProjectDefinition;
+import com.walmartlabs.concord.server.events.ExternalEventResource;
+import com.walmartlabs.concord.server.org.ResourceAccessLevel;
+import com.walmartlabs.concord.server.org.secret.SecretEntry;
 import com.walmartlabs.concord.server.events.Events;
 import com.walmartlabs.concord.server.events.ExternalEventResource;
 import com.walmartlabs.concord.server.events.GithubWebhookService;
@@ -148,7 +151,7 @@ public class ProjectRepositoryManager {
 
     public void validateRepository(UUID projectId, RepositoryEntry repositoryEntry) throws IOException {
         Path srcPath = repositoryManager.fetch(projectId, repositoryEntry);
-        ProjectDefinition pd = loader.load(srcPath);
+        ProjectDefinition pd = loader.loadProject(srcPath);
         repositoryValidator.validate(pd);
     }