+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- project-model: support for multiple Concord project files in
+`concord/*.yml`;
 - concord-server: support for archiving of process state into
 S3-compatible object storages;
 - concord-server: optionally encrypt sensitive data in the process